Adding formal cells now supported
Currently somewhat limited in usage, can only e.g. put an assume on a signal being high, but it is usable.
Moved around child trace propagation to be shared between task types.
Added method to reduce the depth of a task and its children (recursively) for e.g. the add statement.