Fix enabling assumption on the fly
Split commands into pre- and post-sim, adding a post-sim command for chformal to skip activation of the cell in the first cycle where it will still have values from the prior state.
In writing that out there will be issues that arise if the assumption in question can not be cleared in a single cycle, but it works for now.