Changes for local enable cells
i.e. defining an enable/disable in the body of a cover sequence which operates only on that step and not all of the children.
Cleaner way of enabling/disabling for a statement only relevant to a single cover.
Also discovered that one wire can be used for multiple formal cells, so had to do some extra work to disambiguate and preserve all cells that it drives.