Extra updn_cntr tests
Testing (more) valid sequences, particularly those with enable/disable/add statements.
Also checks files are generated (for cover and trace statements) and chunk counts are correct.

Force enable_cells to apply keep to their enable driver so we don't lose it to optimisation in the second step of design_prep.