Handle final trace recovery
Closes #3
Moves final trace to root task, using the same code for both a failed (and recovered) task, and for a successful run.
If `--finaltrace` is being used and an error occurs, stats will still be displayed, with the failed task being marked as such, and any other aborted tasks also marked.
Could potentially move the error handler so that tasks not directly affected by the failed task will still run, but then there could be a situation where multiple branches fail and we would want to be able to trace all of them, so this works for now.
Also better error reporting/handling.